requirements.txt
flask
GeoAlchemy2
networkx
osmnx
psycopg2-binary
SQLAlchemy
